+++ dart/geobase/CHANGELOG.md
@@ -7,6 +7,7 @@ Under development, current pre-release: 0.4.0-dev.1
 🧩 Features:
 * [Add methods to calculate a pixel or a position from tile coordinates in tile matrix sets #158](https://github.com/navibyte/geospatial/issues/158)
 * [Add toText and writeValues on Position and Box #153](https://github.com/navibyte/geospatial/issues/153)
+* [Calculate zoom from pixel resolution or scale denominator #160](https://github.com/navibyte/geospatial/issues/160)
 
 ## 0.3.2